feat: consolidate Vertex AI compatibility with API key support in Gemini Vertex executor
